Criando repository e service para Categoria

+++ src/main/java/com/cursomc/domain/Categoria.java
@@ -15,7 +15,7 @@ public class Categoria implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	private String name;
 	
@@ -23,16 +23,16 @@ public class Categoria implements Serializable{
 		
 	}
 
-	public Categoria(Integer id, String name) {
+	public Categoria(Long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
